requirements.txt
django
djangorestframework
requests
django-background-tasks